refactor localization and particle manager

+++ Convert.h
@@ -16,8 +16,8 @@ public:
 	 * All pixel coordinates to relative robot positions and vice versa don't scale the given values
 	 * Relative robot and resolution should be in meters
 	 */
-	static double PixelXCoordToRobotRelativeXPos(double dX, double resolution, double mapWidth) {
-		double quarterSize = mapWidth / 2;
+	static float PixelXCoordToRobotRelativeXPos(float dX, float resolution, float mapWidth) {
+		float quarterSize = mapWidth / 2;
 		
 		if (dX > quarterSize)
 			dX -= quarterSize;
@@ -29,8 +29,8 @@ public:
 		return dX * resolution;
 	}
 	
-	static double PixelYCoordToRobotRelativeYPos(double dY, double resolution, double mapHeight) {
-		double quarterSize = mapHeight / 2;
+	static float PixelYCoordToRobotRelativeYPos(float dY, float resolution, float mapHeight) {
+		float quarterSize = mapHeight / 2;
 		
 		if (dY > quarterSize) {
 			dY -= quarterSize;
@@ -42,7 +42,7 @@ public:
 		return dY * resolution;
 	}
 	
-	static double RobotRelativeXPosToPixelXCoord(double dX, double resolution, double mapWidth) {
+	static float RobotRelativeXPosToPixelXCoord(float dX, float resolution, float mapWidth) {
 		double quarterSize = (mapWidth / 2) * resolution;
 		
 		if (dX > 0)
@@ -55,8 +55,8 @@ public:
 		return dX / resolution;
 	}
 	
-	static double RobotRelativeYPosToPixelYCoord(double dY, double resolution, double mapHeight) {
-		double quarterSize = (mapHeight / 2) * resolution;
+	static float RobotRelativeYPosToPixelYCoord(float dY, float resolution, float mapHeight) {
+		float quarterSize = (mapHeight / 2) * resolution;
 		
 		if (dY < 0) {
 			dY *= -1;
